requirements.txt
json-tricks==3.15.5
kornia==0.5.2
natsort==7.1.0
numpy==1.19.5
opencv-python==4.5.1.48
PyYAML==5.4.1
scipy==1.6.0
tensorboard==2.4.1
brotlipy==0.7.0
cdflib==0.4.9
chardet==5.1.0
flit-core==3.6.0
h5py==3.8.0
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
notebook==6.4.12
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
pip==22.3.1
pipdeptree==2.9.6
pyOpenSSL==22.0.0
PyQt5-sip==12.11.0
PySocks==1.7.1
pytorch-msssim==0.2.1
sip==6.6.2
torchvision==0.14.1
xarray==0.20.2